+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.4.3 (unreleased)
 
-- Added `add_assumption` and `add_assumptions` to `CpModel`
+- Added `add_assumption`, `add_assumptions`, and `clear_assumptions` to `CpModel`
+- Added `add_hint` and `clear_hints` to `CpModel`
 - Added `only_enforce_if` to `SatConstraint`
 - Fixed installation for Debian